playground-main/requirements.txt
packaging==23.2
pyserde==0.16.0
requests==2.31.0


Pillow==10.0.1
anyio==4.0.0
colorhash==1.2.1
fnvhash==0.1.0
httpx==0.25.0
hypercorn==0.14.4
packaging==23.2
platformdirs==3.11.0
pyperclip==1.8.2
pyserde==0.12.2
requests==2.31.0
sfyi-fsb5==1.1
starlette==0.31.1
typing_extensions==4.8.0
websockets==11.0.3
zstandard==0.21.0